-----The linear chain class has been removed. The structure of the program has been cleaned. Except for N_kink (number of kinks in the circle, which is not used anymore, and anglenums[] which is used to determine the distribution of number of kinks, all other Chain properties, like accepts, moves, endToEndAngle, gyration ratio are moved to Chain.stat.
Test gyration ratio of the following case:
Here are the simulation results for the circular chain of 4 Kuhn segments in length. The value of the gyration radius is expressed in units when Kuhn length=1. This is the result of 10000000 moves.

<rg**2>= 0.2355   err=0.0006

I got 0.2340~0.2347. Not very good match. But not sure if boss has volume exclusion turned on.

I also tested the third momentum of mass distribution. It gives almost zero value, which means the circular chain is manipulated symmetrically. 

Investigation of the mean and variance of each bangle also gives the same result for all bangles. This future reinforce the operations are symmetric.

The project is moving on to construction of rigid body.