-----Added Volume Exclusion. MCbox_circular now reads the config exclusively. Global variables are now not initialized by main program.
Since the MCbox now have to read all variables and then construct dnaChain, the dnaChain cannot be constructed in the parameter list of the constructor. Instead, pointers are used to construct dnaChain *. However this need memory release, which is not added in this version.

REMOVED ALL KINK RELATED MEMBER FUNCTIONS AND VARIALBES.

HOWEVER, THE BANGLE UPDATE FUNCTIONS ARE REMOVED BY MISTAKE. WILL IN ADDED BACK IN VERSION 48.