-----IEV Major Bug Fixed: 		//NOTE ABOUT M AND N. PAY ATTENTION. 		//M<N is not necessary here since deltaE_TrialCrankshaft_countMove and crankshaft 		//support M>N and automatically wraps the iterator when it hits the tail to the 		//beginning of the chain. 		//Therefore, all energy evaluation program should be careful with chain segment 		//iteration due to wrapping problem.
The IEV function used to assume m<n and compare chain from m~n with the rest of the body. However m>n could happen and wrapping around from tail to head is available in other functions. Therefore, if m and n are not swapped in IEV, NO CHAIN segments will be checked for IEV.

Some variables are also moved to config file for initialization.