-----CircularChain::checkConsistancy() added to check if X[i+1]-X[i]==dX[i] (including wrap around). CircularChain::updateKPoly and CircularChain::AlexPoly added to facilitate the calculation and enhance encapsulation.
Big lesson learned that if the input conformation is not consistent (i.e. not normalized, circle is not closed etc, the program won't discover it, until reptation movement bring that inconsistency to other segment inside the chain, not at the end of the chain.

The repation movement is tested by mannually setting:
	for (int i=0;i<40;i++){
		dnaChain->C[i].x=i*(i-1)/2;dnaChain->C[i].y=0;dnaChain->C[i].z=0;
		dnaChain->C[i].dx=i;dnaChain->C[i].dy=0;dnaChain->C[i].dz=0;
	}
	dnaChain->snapshot("1.txt");
	dnaChain->dE_reptation(1,32,-1);
	dnaChain->snapshot("2.txt");
	dnaChain->dE_reptation(1,32,1);
	dnaChain->snapshot("3.txt");
	exit(0);

It works very well. And reptation + move and repation -move can bring the conformation back without any descrapancy.