-----There is still problem with CircularChain::kpoly. Some confirmation (Like this initial conformation file main.vmc) is calculated as Kpoly(1,1) (t=-1, and t=-2). Which is incorrect. (as shown in main_log.txt)
However _kndwr() function provides a topl return value that calculate t=-1 (?) topl>1.5 there is knot. Otherwise there is no knot. It is very fast. And meanwhile KNDWR (directional writhe number) can be calculated.

Using this topl as a flag to see if there is a knot, the main.vmc will generate no move at all since it is actually a 71 knot. This proves topl works better than kpoly, but doesn't mean that topl and _kndwr are correct. More tests needed.

The snapshot function is also revered back to the "imperfect" old version which snapshot the circle as a linear chain, with one more bead added. This actually facilitate convertion "convertJmol2For.py" which is designed to convert Jmol file for linear chain.