-------Artificial vertex rigidity. Set arbitrary rigidity to assigned vertex.
In this version, in order to test the importance of the direction of the vectors that connects the rigid core to the surrounding chain, rigidity of the vertices at the boundary of the synaptic core is largely or completely removed.

flexcore.vmc is directly from 20100819_Flexcore_r108, not from 20100901_SiteIonly_r109.