-------Successfully implemented two res site with heterogeneous segmentation, however the reptation movements hasn't been tested yet.
Added RBAUS to r to smooth the distribution of r (siteII-siteII distance) with constraint that when r<7, the AxisBeta and RadiusBeta are within the range of 60 degrees. By sampling using RBAUS for 0<r<50 and refined sampling within 0<r<10, a nice ArtificialPotential has been obtained to smooth the distribution of r.