-----in bias.h, stringstream(buf)>>buf doesn't compile on Intel C++ compiler. MSVC is non-standard and could compile this segment.
MSVC compiler is non-standard. Intel compiler abides by the C++98 or 03 standard and the reason of the problem is explained in the following post on VelocityReviews website:

The operator>> is a non-member. It takes the first argument by a
non-const reference, which cannot be initialized with a temporary.
Define a named variable and you will be able to do what you want.